Initial commit message
Changed the html and css file names. Added the links for the css files. Created the ws_captions.vtt file for the captions in the video. Added the video clip and the possible sources for the video clip. Added the paragraph if browser not supported. Added the media quarry for the screen with min 521 pixels. Added text and background css styles for the captions. Added video styles for the video clip.